+++ src/Elements/Geometry/Profile.cs
@@ -19,7 +19,10 @@ namespace Elements.Geometry
         }
 
         /// <summary>
-        /// Construct a profile from a list of polygons. Inner polygons will be treated as voids.
+        /// Construct a profile from a collection of polygons.
+        /// If the collection contains more than one polygon, the first polygon
+        /// will be used as the perimeter and any remaining polygons will 
+        /// be used as voids.
         /// </summary>
         /// <param name="polygons">The polygons bounding this profile.</param>
         public Profile(IList<Polygon> polygons) : base(Guid.NewGuid(), null)
@@ -57,7 +60,7 @@ namespace Elements.Geometry
             {
                 throw new ArgumentException("Unable to construct a profile. More than one of the polygons supplied are not contained by any other.");
             }
-            if(outerMostIndices.Count() == 0)
+            if (outerMostIndices.Count() == 0)
             {
                 throw new ArgumentException("Unable to construct a profile. All the supplied polygons are inside other supplied polygons. Sounds like a geometric paradox!");
             }
@@ -73,10 +76,10 @@ namespace Elements.Geometry
         /// Construct a profile.
         /// </summary>
         /// <param name="perimeter">The perimeter of the profile.</param>
-        /// <param name="singleVoid">A void in the profile.</param>
+        /// <param name="void">A void in the profile.</param>
         public Profile(Polygon perimeter,
-                       Polygon singleVoid) :
-            this(perimeter, new[] { singleVoid }, Guid.NewGuid(), null)
+                       Polygon @void) :
+            this(perimeter, new[] { @void }, Guid.NewGuid(), null)
         { }
 
         /// <summary>
@@ -121,6 +124,7 @@ namespace Elements.Geometry
                 this.Voids[i].Transform(t);
             }
         }
+
         /// <summary>
         /// Return a new profile that is this profile scaled about the origin by the desired amount.
         /// </summary>